+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 - Update the WhatsApp integration to use Twilio Content Templates
 - Add auto-refresh functionality to the Log page (#957, @mickBoat00)
 - Redesign the "Status Badges" page
+- Add support for per-check status badges (#853)
 
 ### Bug Fixes
 - Fix Gotify integration to handle Gotify server URLs with paths (#964)